+++ packages/system/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @marigold/system
 
+## 7.5.1
+
+### Patch Changes
+
+- [#3809](https://github.com/marigold-ui/marigold/pull/3809) [`fba5e92`](https://github.com/marigold-ui/marigold/commit/fba5e92e8e70de0c3a65696cd2c462b6dfa7819e) Thanks [@OsamaAbdellateef](https://github.com/OsamaAbdellateef)! - Preventing to override our theme variants in extendTheme
+
 ## 7.5.0
 
 ### Minor Changes